pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "icalendar>=6.3.2",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "tavily-python>=0.7.17",
]
